https://github.com/paypal/squbs/issues/437 Changed RequestContext id and generator to Long.

+++ squbs-pipeline/src/main/scala/org/squbs/pipeline/RequestContext.scala
@@ -22,7 +22,7 @@ import scala.collection.JavaConversions._
 import scala.util.{Success, Try}
 
 case class RequestContext(request: HttpRequest,
-                          httpPipeliningOrder: Int, // TODO Come up with a better val name
+                          httpPipeliningOrder: Long, // TODO Come up with a better val name
                           response: Option[Try[HttpResponse]] = None,
                           attributes: Map[String, Any] = Map.empty) {
